port drawing
I fixed the code so that ports are not drawn on the open ocean.

Currently in map_generation.resetGrid I gave a constant seed to get the same map every time because I need to fix the port drawing issue. The problem is that there's a difference between where a road enters the water (should have a port) and where ports are drawn by map_generation.connectByRoad.
I want to fix it so that ports aren't strung along the coast.